+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### Added
 - The Architect Table in the Garrison will now automatically designate a plot for the Barracks, provided that the player's Garrison is at level 1 and they are currently on the "Build Your Barracks" quest.
+- The Mission Table in the Garrison will now have its first mission automated, provided the player has yet to send their first mission and the player's Garrison is level 1.
 ### Changed or Updated
 ### Deprecated
 ### Removed